chore(ui): rename components for organizational clarity
