+++ ui/src/CloneManagerCreateDialog.svelte
