+++ ui/src/CloneManagerShareDialog.svelte
